(refactored) the code to change status.
The code display whenever a child is changed, in our case the status key
(Online/Offline). But we need make it do it as soon as someone logs in,
now it only does it when you actually change it in the db. Also need to
figure out how to remove a user object from the array once the user goes
offline.